refactor: move types to typings/ folder
