+++ bootcamp-milestone-2/src/app/blogData.ts
@@ -1,11 +1,4 @@
-export interface Blog {
-  title: string;
-  date: string;
-  description: string;
-  image: string;
-  imageAlt: string;
-  slug: string;
-}
+import type { Blog } from "@/typings/blog";
 
 const blogs: Blog[] = [
   {
