+++ bootcamp-milestone-2/src/typings/blog.ts
@@ -1,0 +1,8 @@
+export interface Blog {
+  title: string;
+  date: string;
+  description: string;
+  image: string;
+  imageAlt: string;
+  slug: string;
+}
