+++ bootcamp-milestone-2/src/typings/project.ts
@@ -1,0 +1,7 @@
+export interface Project {
+  title: string;
+  description: string;
+  image: string;
+  imageAlt: string;
+  url: string;
+}
